Move OffenseTable components to its own directory
